More x64 compilation fixes

+++ kauai/src/util.h
@@ -60,8 +60,8 @@ typedef HCURSOR HCRS;
 
 #endif // WIN
 
-#define size(foo) ((long)sizeof(foo))
-#define offset(FOO, field) ((int)&((FOO *)0)->field)
+#define size(foo) ((size_t)sizeof(foo))
+#define offset(FOO, field) ((size_t)&((FOO *)0)->field)
 #define CvFromRgv(rgv) (size(rgv) / size(rgv[0]))
 #define BLOCK
